Refactor: initial change of `VoteResults`
