+++ src/modules/dashboard/components/ActionsPage/FinalizeMotionWidget/VoteResults/VoteResults.css
@@ -1,95 +1,8 @@
-.wrapper {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  max-width: 320px;
+.main {
+  display: block;
+  margin-top: 5px;
 }
 
-.voteInfoContainer {
-  display: flex;
-  align-items: flex-end;
-  width: 85%;
-}
-
-.voterAvatarsContainer {
-  position: relative;
-}
-
-.voteResults {
-  display: flex;
-  flex-direction: column;
-  margin-left: 10px;
-  width: calc(100% - 38px);
-}
-
-.voteHeading {
-  display: flex;
-  align-items: center;
-  margin-bottom: 5px;
-}
-
-.votePercentageBar {
-  height: 6px;
-  border-radius: 4px;
-}
-
-.themeApprove .votePercentageBar {
-  background-color: var(--primary);
-}
-
-.themeDisapprove .votePercentageBar {
-  background-color: var(--pink);
-}
-
-.voterAvatars {
-  display: flex;
-  margin-left: -24px;
-  position: relative;
-}
-
-.voterAvatar {
-  position: relative;
-  border-radius: 50px;
-  border: 1px solid var(--colony-white);
-}
-
-.voterAvatar:first-child {
-  left: 24px;
-  z-index: 2;
-}
-
-.voterAvatar:nth-child(2),
-.voterAvatar:nth-last-child(2) {
-  left: 12px;
-  z-index: 1;
-}
-
-.voterAvatar:nth-child(3),
-.voterAvatar:last-child {
-  left: 0;
-  z-index: 0;
-}
-
-.remaningAvatars {
-  margin: 0;
-  width: 22px;
-  position: absolute;
-  top: 0;
-  right: -12px;
-  vertical-align: middle;
-  z-index: 3;
-  border-radius: var(--radius-large);
-  background-color: var(--grey-blue-2);
-  font-size: 9px;
-  font-weight: var(--weight-bold);
-  text-align: center;
-  line-height: 14px;
-  color: var(--temp-grey-blue-7);
-}
-
-.votePercentage {
-  margin-left: 5px;
-  font-size: var(--size-tiny);
-  font-weight: var(--weight-bold);
-  color: var(--temp-grey-blue-7);
+.firstVoteResult {
+  margin-bottom: 20px;
 }
