+++ src/modules/dashboard/components/ActionsPage/FinalizeMotionWidget/VoteResults/VoteResults.css.d.ts
@@ -1,12 +1,2 @@
-export const wrapper: string;
-export const voteInfoContainer: string;
-export const voterAvatarsContainer: string;
-export const voteResults: string;
-export const voteHeading: string;
-export const votePercentageBar: string;
-export const themeApprove: string;
-export const themeDisapprove: string;
-export const voterAvatars: string;
-export const voterAvatar: string;
-export const remaningAvatars: string;
-export const votePercentage: string;
+export const main: string;
+export const firstVoteResult: string;
