+++ src/modules/dashboard/components/ActionsPage/FinalizeMotionWidget/VoteResults/VoteResults.tsx
@@ -1,99 +1,52 @@
-import React from 'react';
-import { MessageDescriptor } from 'react-intl';
+import React, { useCallback } from 'react';
+import { defineMessages, FormattedMessage } from 'react-intl';
 
-import Icon from '~core/Icon';
-import Heading from '~core/Heading';
-import ProgressBar from '~core/ProgressBar';
-import HookedUserAvatar from '~users/HookedUserAvatar';
-import { Address } from '~types/index';
-import { getMainClasses } from '~utils/css';
-import useAvatarDisplayCounter from '~utils/hooks/useAvatarDisplayCounter';
+import { Colony, useLoggedInUser } from '~data/index';
+import { ActionTypes } from '~redux/index';
+import { ColonyMotions } from '~types/index';
+
+import VoteResultsItem from './VoteResultsItem';
 
 import styles from './VoteResults.css';
 
-interface Appearance {
-  theme?: 'approve' | 'disapprove';
-}
-
 interface Props {
-  appearance?: Appearance;
-  value: number;
-  maxValue: number;
-  maxPercentage?: number;
-  title: MessageDescriptor;
-  voters?: Address[];
-  maxAvatars?: number;
+  colony: Colony;
+  motionId: number;
 }
 
-const displayName = 'dashboard.ActionPage.FinalizeVoteWidget.VoteResults';
-
-const VoteResults = ({
-  appearance = { theme: 'approve' },
-  value,
-  maxValue,
-  maxPercentage = 100,
-  title,
-  voters = [],
-  maxAvatars = 3,
-}: Props) => {
-  const {
-    avatarsDisplaySplitRules,
-    remainingAvatarsCount,
-  } = useAvatarDisplayCounter(maxAvatars, voters);
-  const UserAvatar = HookedUserAvatar({ fetchUser: true });
-  const iconName =
-    appearance.theme === 'approve' ? 'circle-thumbs-up' : 'circle-thumbs-down';
-  const votePercentage = (value * maxPercentage) / maxValue;
-  const barTheme = appearance.theme === 'approve' ? 'primary' : 'danger';
-
+const VoteResults = ({ colony: { colonyAddress }, colony }: Props) => {
   return (
-    <div className={`${styles.wrapper} ${getMainClasses(appearance, styles)}`}>
-      <div className={styles.voteInfoContainer}>
-        <Icon name={iconName} title={title} appearance={{ size: 'medium' }} />
-        <div className={styles.voteResults}>
-          <div className={styles.voteHeading}>
-            <Heading
-              appearance={{
-                theme: 'dark',
-                size: 'small',
-                weight: 'bold',
-                margin: 'none',
-              }}
-              text={title}
-            />
-            <span className={styles.votePercentage}>{votePercentage}%</span>
-          </div>
-          <ProgressBar
-            value={votePercentage}
-            max={maxPercentage}
-            appearance={{
-              size: 'small',
-              backgroundTheme: 'transparent',
-              barTheme,
-            }}
-          />
-        </div>
+    <div className={styles.main}>
+      <div className={styles.firstVoteResult}>
+        <VoteResultsItem
+          value={20}
+          maxValue={100}
+          title="yes"
+          voters={[
+            '0xb77D57F4959eAfA0339424b83FcFaf9c15407461',
+            '0x9dF24e73f40b2a911Eb254A8825103723E13209C',
+          ]}
+        />
       </div>
-      <div className={styles.voterAvatarsContainer}>
-        <ul className={styles.voterAvatars}>
-          {voters
-            .slice(0, avatarsDisplaySplitRules)
-            .map((voterAddress: Address) => (
-              <li className={styles.voterAvatar} key={voterAddress}>
-                <UserAvatar size="xs" address={voterAddress} notSet={false} />
-              </li>
-            ))}
-        </ul>
-        {!!remainingAvatarsCount && (
-          <span className={styles.remaningAvatars}>
-            {remainingAvatarsCount < 99 ? `+${remainingAvatarsCount}` : `+99`}
-          </span>
-        )}
-      </div>
+      <VoteResultsItem
+        value={80}
+        maxValue={100}
+        title="no"
+        appearance={{ theme: 'disapprove' }}
+        voters={[
+          '0xb77D57F4959eAfA0339424b83FcFaf9c15407461',
+          '0x9dF24e73f40b2a911Eb254A8825103723E13209C',
+          '0xb77D57F4959eAfA0339424b83FcFaf9c15407461',
+          '0xb77D57F4959eAfA0339424b83FcFaf9c15407461',
+          '0x9dF24e73f40b2a911Eb254A8825103723E13209C',
+          '0xb77D57F4959eAfA0339424b83FcFaf9c15407461',
+          '0x9dF24e73f40b2a911Eb254A8825103723E13209C',
+        ]}
+      />
     </div>
   );
 };
 
-VoteResults.displayName = displayName;
+VoteResults.displayName = 'dashboard.ActionsPage.VoteResults';
 
 export default VoteResults;
