+++ src/modules/dashboard/components/ActionsPage/FinalizeMotionWidget/VoteResults/VoteResultsItem.css.d.ts
@@ -1,0 +1,12 @@
+export const wrapper: string;
+export const voteInfoContainer: string;
+export const voterAvatarsContainer: string;
+export const voteResults: string;
+export const voteHeading: string;
+export const votePercentageBar: string;
+export const themeApprove: string;
+export const themeDisapprove: string;
+export const voterAvatars: string;
+export const voterAvatar: string;
+export const remaningAvatars: string;
+export const votePercentage: string;
